% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bigfour_2016}
\alias{bigfour_2016}
\title{Big four major sports from 2016}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 32 rows and 6 columns.}
\usage{
bigfour_2016
}
\description{
Big four major sports from 2016
}
\keyword{datasets}
