% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bigfour_end}
\alias{bigfour_end}
\alias{best16_2016}
\alias{spurs_bulls}
\title{End-of-season parameters for four major sports}
\format{\describe{
 \item{sport}{which sport}
 \item{season}{which season, 1 = 2004}
 \item{team_id}{id of the team within each sport}
 \item{name}{team name}
 \item{N}{number of observations aggregated}
 \item{mean_theta}{average value of theta (team strength) over the final
 \code{N} weeks of the \code{season}}
 \item{mean_alpha}{average value of alpha (home advantage)}
 \item{alpha_sport}{sport-specific home advantage}
}}
\usage{
bigfour_end

best16_2016

spurs_bulls
}
\description{
End-of-season parameters for four major sports

Best 16 teams in each sport for 2016

Spurs vs. Bulls in 2016
}
\keyword{datasets}
