% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.R
\name{finish_tournament}
\alias{finish_tournament}
\title{Finalize tournament results}
\usage{
finish_tournament(g, ...)
}
\arguments{
\item{g}{a tournament}

\item{...}{currently ignored}
}
\description{
Finalize tournament results
}
