% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{home_advantages}
\alias{home_advantages}
\title{Home advantage estimates}
\format{\describe{
 \item{alpha.team}{team-specific home advantage estimate}
 \item{alpha.lower}{5\% bound for team-specific home advantage estimate}
 \item{alpha.upper}{95\% bound for team-specific home advantage estimate}
 \item{team}{team name}
 \item{sport}{which sport}
 \item{alpha.sport}{sport-specific home advantage estimate}
 \item{alpha.team.overall}{sum of \code{alpha.team} and \code{alpha.sport}}
 \item{alpha.team.lower}{5\% bound for team-specific total home advantage estimate}
 \item{alpha.team.upper}{95\% bound for team-specific total home advantage estimate}
 \item{team_id}{id of the team within each sport}
}}
\usage{
home_advantages
}
\description{
Home advantage estimates
}
\keyword{datasets}
