% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.R
\name{one_simulation}
\alias{one_simulation}
\alias{many_simulations}
\title{Simulate a single tournament}
\usage{
one_simulation(data, series_length = 1, ...)

many_simulations(data, n = 2, series_length = 1, ...)
}
\arguments{
\item{data}{a data frame}

\item{series_length}{number of games in each series}

\item{...}{arguments passed to \code{\link{play}}}

\item{n}{number of simulations}
}
\description{
Simulate a single tournament

Simulate many tournaments
}
\examples{

n <- 8
teams <- data.frame(sport = "nba", id = 1:n,
                    mean_theta = sort(runif(n), decreasing = TRUE))

one_simulation(teams)
one_simulation(teams, series_length = 7)
one_simulation(teams, series_length = 99)

res <- many_simulations(teams, n = 10, series_length = 1)
res <- many_simulations(teams, n = 10, series_length = 7)

\dontrun{
  res <- many_simulations(teams, n = 100, series_length = 7)
  with(res, cor(seed, finish))
}

}
