% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.R
\name{play}
\alias{play}
\title{Play simulated tournament games}
\usage{
play(g, node = 1, series_length = 1, ...)
}
\arguments{
\item{g}{a tournament}

\item{node}{index of node to play}

\item{series_length}{length of each series}

\item{...}{currently ignored}
}
\description{
Play simulated tournament games
}
