% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.R
\name{seed_tournament}
\alias{seed_tournament}
\title{Seed a tournament}
\usage{
seed_tournament(data, ...)
}
\arguments{
\item{data}{a data frame}

\item{...}{currently ignored}
}
\description{
Seed a tournament
}
