% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.R
\name{series_probability}
\alias{series_probability}
\title{Calculate the probability of winning a series}
\usage{
series_probability(theta1, theta2, series_length = 1, ...)
}
\arguments{
\item{theta1}{value of theta}

\item{theta2}{value of theta for other team}

\item{series_length}{number of games in the series}

\item{...}{currently ignored}
}
\description{
Calculate the probability of winning a series
}
\examples{
if (require(dplyr)) {
  spurs_bulls <- bigfour_2016 \%>\%
    filter(sport == "nba", season == 10,
           team_id \%in\% c(27, 5))

  # should be around 54\%
  series_probability(spurs_bulls$mean_theta[1], spurs_bulls$mean_theta[2], 1)
  long_run_prob(spurs_bulls, n = 10)

  series_probability(spurs_bulls$mean_theta[1], spurs_bulls$mean_theta[2], 7)
  long_run_prob(spurs_bulls, n = 10, series_length = 7)

  # should be nearly certain
  series_probability(spurs_bulls$mean_theta[1], spurs_bulls$mean_theta[2], 99)
  long_run_prob(spurs_bulls, n = 10, series_length = 99)
}
}
