% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{team_strengths}
\alias{team_strengths}
\title{Team strength estimates}
\format{\describe{
 \item{theta}{team strength estimate}
 \item{season}{which season, 1 = 2004}
 \item{week}{week within each season}
 \item{team_id}{id of the team within each sport}
 \item{cumweek}{cumulative weeks for each team across all seasons}
 \item{sport}{which sport}
 \item{max.week}{maximum number of weeks in each season, in each sport}
 \item{time_val}{numerical measure of time}
 \item{name}{team name}
}}
\usage{
team_strengths
}
\description{
Team strength estimates
}
\keyword{datasets}
