% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.R
\name{tournament_ordering}
\alias{tournament_ordering}
\title{Set tournament ordering}
\usage{
tournament_ordering(x)
}
\arguments{
\item{x}{a vector of potential seeds}
}
\description{
Set tournament ordering
}
\examples{
tournament_ordering(1:16)
}
