% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tournament.R
\name{one_simulation}
\alias{one_simulation}
\alias{many_simulations}
\title{Simulate a single tournament}
\usage{
one_simulation(data, series_length = 1, ...)

many_simulations(data, n = 2, series_length = 1, ...)
}
\arguments{
\item{data}{a data frame}

\item{series_length}{number of games in each series}

\item{...}{arguments passed to \code{\link{play}}}

\item{n}{number of simulations}
}
\description{
Simulate a single tournament

Simulate many tournaments
}
\examples{

if (require(dplyr)) {
nba <- bigfour_2016 \%>\%
  filter(sport == "nba")

one_simulation(nba)
one_simulation(nba, series_length = 7)
one_simulation(nba, series_length = 99)

res <- bigfour_2016 \%>\%
  group_by(sport) \%>\%
  do(many_simulations(., n = 10, series_length = 7))
res \%>\%
  group_by(sport) \%>\%
  summarize(cor(seed, finish))
}

}
